attached_assets/requirements_1758097345919.txt
bcrypt>=4.3.0
pandas>=2.3.2
pdfplumber>=0.11.7
plotly>=6.3.0
pypdf2>=3.0.1
streamlit>=1.49.1
google-generativeai
python-dotenv